01_ingest_layer/requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
piexif>=1.1.3
tqdm>=4.66.0
